FIX: Errors Moved back to Flask App
- Blueprints not working as I had hoped with errors, so I moved back
to the main app for now to get things working.
